pdf-summarizer-map-reduce-langchain/BE/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
openai
pypdf
python-dotenv
python-multipart